+++ changelog.rst
@@ -7,6 +7,7 @@ next release
 * Adding missing dev packages on docker images.
 * Docker images clone latest branch of pwnypack repository.
 * Fix magic command registration on IPython >= 5.0.0.
+* Support parsing (a subset of) ~/.ssh/config in flow.SSHClient.
 
 0.8.0 (2015-05-17)
 ==================